@@ -1,7 +1,7 @@
 package com.xiaopo.flying.photolayout
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 
 class CreateQuoterActivity: AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
